requirements.txt
# Minimal dependencies for Vercel serverless deployment
# For full local development, use requirements.full.txt

# OpenAI API Client (essential for chatbot)
openai>=1.50.0

# Web API (essential)
fastapi>=0.111.0

# Data Validation
pydantic>=2.0.0

# Environment Configuration
python-dotenv>=1.0.0
